Refactor layout: replace tabnav under project with a project actions section on the main page. Convert all individual pages to sheets and dialogs triggered by project actions cards.
