--- src/routes/projects.$project_id.actions.route.tsx
@@ -1,15 +1,0 @@
-import { Outlet, createFileRoute } from '@tanstack/react-router'
-
-export const RouteComponent = () => (
-  <>
-    <Outlet />
-  </>
-)
-
-export const Route = createFileRoute('/projects/$project_id/actions')({
-  component: RouteComponent,
-  loader: () => ({
-    crumb: 'Project Actions',
-    includeCrumbLink: true,
-  }),
-})
