--- src/routes/projects.$project_id.files.route.tsx
@@ -1,15 +1,0 @@
-import { Outlet, createFileRoute } from '@tanstack/react-router'
-
-export const RouteComponent = () => (
-  <>
-    <Outlet />
-  </>
-)
-
-export const Route = createFileRoute('/projects/$project_id/files')({
-  component: RouteComponent,
-  loader: () => ({
-    crumb: 'Files',
-    includeCrumbLink: true,
-  }),
-})
