+++ src/routes/projects.$project_id.route.tsx
@@ -1,8 +1,6 @@
 import { AxiosError } from 'axios'
 import { Outlet, createFileRoute, getRouteApi, redirect } from '@tanstack/react-router'
-import { FolderOpen, HardDriveDownload, LayoutDashboard, Zap } from 'lucide-react'
 import { getProjectByProjectId } from '@/client'
-import { TabLink, TabNav } from '@/components/tab-nav'
 
 export const Route = createFileRoute('/projects/$project_id')({
   component: RouteComponent,
@@ -32,38 +30,9 @@ function RouteComponent() {
   return (
     <>
       <div className='flex flex-col gap-4'>
-        {/* Header and tab navigation */}
+        {/* Header */}
         <h1 className='text-3xl font-extralight'>{project.name}</h1>
-        <div className='flex gap-2 flex-col md:flex-row md:h-9 md:items-center'>
-          <TabNav>
-            <TabLink
-              to='/projects/$project_id/overview'
-              params={{ project_id: project.project_id }}
-            >
-              <LayoutDashboard /><span>Overview</span>
-            </TabLink>
-            <TabLink
-              to='/projects/$project_id/files'
-              params={{ project_id: project.project_id }}
-            >
-              <FolderOpen /><span>Files</span>
-            </TabLink>
-            <TabLink
-              to='/projects/$project_id/ingest'
-              params={{ project_id: project.project_id }}
-            >
-              <HardDriveDownload /> <span>Ingest Vendor Data</span>
-            </TabLink>
-            <TabLink
-              to='/projects/$project_id/actions'
-              params={{ project_id: project.project_id }}
-            >
-              <Zap /> <span>Project Actions</span>
-            </TabLink>            
-          </TabNav>
-          
-        </div>
-        {/* Tab nav outlet */}
+        {/* Outlet */}
         <Outlet />
       </div>
     </>
